Integrate Alex's RBTree containment cache implementation with the current code base For now I just copied and pasted files because I was lazy: we may want to create a library out of Alex's code and publish a version to the our eear-maven repo and pull from there; we can discuss this later.
Make the UNSAT and SAT ContainmentCacheEntries implement ICacheEntry<Station>
Extract the relevant parts of ContainmentCache.java into a new class, ContainmentCacheBUndle (terrible name, should change) which essentially wraps an UNSAT and SAT ContainmentCache object and performs the additional checks required. This interface was made to match what the ContainmentCacheController (the web interface) was using

I added an addAll method to the ContainmentCache
I changed getSuper/Subsets to return an Iterator instead of a HashSet because we want short circuiting operations. It remains to be verified that I've done this correctly

This is still untested, and will need to be compared (performance wise) against our previous implementation)