Clasp3SATSolver class previously contained a field which was a scheduled executor that waited for 5 minutes past the cutoff time and forced a Sys.exit() if the job hadn't completed
This posed 2 problems:
1) People got pissed off at us for calling System.exit(), since we are a library and that's bad practice
2) We create TONS of these Clasp3SATSolver objects, and each one had a thread pool, which means we were using up threads for no reason. They weren't harmful (they probably just slept most of the time), but they still count toward os resource limits (e.g. ulimit -u). This could have been solved with a static variable, but then we'd need a similar mechanism for DCCA or whatever else comes next. So...

Borrowing from a StackOveflow answer, we now simply have a timer surrounding the entire call to SolverBundle#solve, which will throw an exception if the cutoff is not obeyed. The bad news is that this conveys a bit less information (e.g. before, you knew that all the threads were definitely stuck in clasp). But the logs should reveal that anyways.

Closes #132
Closes #131