Fix bug in clasp JNA bridge where we are reading values from the symbol table instead of the model
This makes a difference if preprocessing is used!

In particular, in the line:
assignment_[i] = (s.value(v) == value_false ? -1:1)*v;
you are using s.value(v) instead of m.value(v).
The latter would give you the value of the variable v
in the current model. The former returns either the value of v
in the solver's assignment or value_true if v was eliminated by
the preprocessor.